refactor: address PR review for env-module transform
