+++ examples/with-solid-start/src/global.d.ts
@@ -1,4 +1,1 @@
 /// <reference types="@solidjs/start/env" />
-
-// Fullstack apps use `import { env } from "./env"` (transform mode).
-// SPA mode can still augment import.meta.env via ImportMetaEnvAugmented.
