+++ packages/vite-plugin/tsdown.config.ts
@@ -7,6 +7,9 @@ export default defineConfig({
 	fixedExtension: false,
 	deps: {
 		alwaysBundle: ["@repo/log", "@repo/types", "@repo/utils"],
+		// jiti loads the user's env.ts at build time; keep it external so the
+		// plugin does not ship a second copy and Node can resolve project deps.
+		// @arkenv/build is a workspace runtime dependency for key extraction.
 		neverBundle: ["vite", "@arkenv/build", "jiti"],
 	},
 });
